gateway-service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pyjwt
